requirements.txt
temporalio>=1.12.0
GitPython>=3.1.44
requests>=2.30.0
